# Use ArrayPool for EventProperty arrays