No changes in public API.