@nblumhardt 2023-01-24
I think we should reconsider this one, unless the benchmark differences are stark. Adam Sitnik's investigation suggests that `new` + GC will be faster for small arrays, as I'd expect most of these will be.

https://adamsitnik.com/Array-Pool/#the-results

We'd also need to carefully consider where exceptions may be thrown between the rent and return phases, since we'll risk leaking space in the pool otherwise. Adding `finally` blocks would mitigate this but introduces more noise and complexity that the benchmark improvements would need to counterbalance.