@sungam3r 2023-01-25
Log line changed to ` _log.Information(_exception, "Hello, {Name} {Name2} {Name3} {Name4}!", "World", "World2", "World3", "World4");`:

before

|       Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| EmitLogEvent | 262.5 ns | 5.20 ns | 5.34 ns | 0.0753 |     632 B |

after 

|       Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
|------------- |---------:|--------:|--------:|-------:|----------:|
| EmitLogEvent | 258.5 ns | 2.47 ns | 2.31 ns | 0.0648 |     544 B |