@sungam3r 2023-01-25
1. 15% slower, 10% less memory
2. 1,5% slower, 14% less memory

Of course, experience can vary.