@sungam3r 2023-01-25
Honestly, I would prefer to refactor the way Serilog works with `EventProperty` and other internal/public stuff but it's not just an optimization, it also requires changes in public APIs. The effect can be an avalanche-like. 

In this regard Serilog is somewhat similar to Newtonsoft.Json (compared to STJ). It provides rich possibilities and object model at the cost of significant memory consumption for the objects created.