@nblumhardt 2023-01-25
A couple of years ago I spent some time working on a prototype for a version of Serilog based on struct log events and a matching "middleware"-ish style of pipeline using ref parameters. To avoid forking the ecosystem, the idea was to support both pipeline styles, but only fall back to allocating `LogEvent` objects for the old pipeline when previous-version components (sinks and so on) were present.

On the plus side, I think this could work, but on the minus, it involves quite a lot of machinery and corresponding development effort. Could still be something to explore, though! :-)

On this PR, I think we're better off leaving things as-is - the exception-safety problems around array pooling outweigh the benefits of lower allocation.