@nblumhardt 2023-02-22
Looping back - I was wrong about exception safety being a worry, here - it's fine to leak rented arrays. I'll zip back over, fix the conflict, and merge this when I have a chance 👍