@sungam3r 2023-02-23
👍 Glad to hear it but... now I doubt whether it really needs to be done 😄. As I said previously
> Honestly, I would prefer to refactor the way Serilog works with EventProperty and other internal/public stuff but it's not just an optimization, it also requires changes in public APIs. 

If you are sure with suggested changes here then OK, I'm fine merging it and can resolve conflict on my own. Ping me back for your decision.