@nblumhardt 2023-02-25
Thanks for the reply; sorry about the very slow turnaround at the moment, currently very deep in another project so limited cycles to spare for a week or two more.

I might dig a bit more into performance as soon as I have a chance 👍