@nblumhardt 2023-03-29
I think this could be one to skip, given that small arrays will make up a disproportionate amount of events in typical apps.

@sungam3r I'll close this since I have a day to run over and triage the open PRs - trying to get back to a nice clean tracker. Feel free to ping/reopen if you think we should still be exploring this one. Thanks for all the effort on it, though!